# internal/interfaces/http/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/budgets": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Lista todos os orçamentos do usuário com status atualizado",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "budgets"
                ],
                "summary": "List budgets",
                "responses": {
                    "200": {
                        "description": "Lista de orçamentos",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_vasconcellos_financial-control_internal_domain_dto.BudgetResponse"
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Cria um orçamento para controlar gastos por categoria",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "budgets"
                ],
                "summary": "Create a budget",
                "parameters": [
                    {
                        "description": "Dados do orçamento",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_vasconcellos_financial-control_internal_domain_dto.CreateBudgetRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Orçamento criado",
                        "schema": {
                            "$ref": "#/definitions/github_com_vasconcellos_financial-control_internal_domain_dto.BudgetResponse"
                        }
                    },
                    "400": {
                        "description": "Dados inválidos",
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/categories": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Lista todas as categorias do usuário",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "categories"
                ],
                "summary": "List categories",
                "responses": {
                    "200": {
                        "description": "Lista de categorias",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_vasconcellos_financial-control_internal_domain_dto.CategoryResponse"
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Cria uma nova categoria de transação (income ou expense)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "categories"
                ],
                "summary": "Create a new category",
                "parameters": [
                    {
                        "description": "Dados da categoria",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_vasconcellos_financial-control_internal_domain_dto.CreateCategoryRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Categoria criada",
                        "schema": {
                            "$ref": "#/definitions/github_com_vasconcellos_financial-control_internal_domain_dto.CategoryResponse"
                        }
                    },
                    "400": {
                        "description": "Dados inválidos",
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/categories/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Remove uma categoria",
                "tags": [
                    "categories"
                ],
                "summary": "Delete a category",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID da categoria",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Categoria removida"
                    },
                    "401": {
                        "description": "Não autenticado",
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Categoria não encontrada",
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/goals": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Lista todas as metas financeiras do usuário",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "goals"
                ],
                "summary": "List goals",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Número máximo de resultados",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset para paginação",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de metas",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_vasconcellos_financial-control_internal_domain_dto.GoalResponse"
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Cria uma nova meta financeira (ex: guardar R$ 10.000)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "goals"
                ],
                "summary": "Create a financial goal",
                "parameters": [
                    {
                        "description": "Dados da meta",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_vasconcellos_financial-control_internal_domain_dto.CreateGoalRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Meta criada",
                        "schema": {
                            "$ref": "#/definitions/github_com_vasconcellos_financial-control_internal_domain_dto.GoalResponse"
                        }
                    },
                    "400": {
                        "description": "Dados inválidos",
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_vasconcellos_financial-control_internal_domain_dto.BudgetResponse": {
            "type": "object",
            "properties": {
                "alertPercent": {
                    "type": "number"
                },
                "amount": {
                    "type": "number"
                },
                "categoryId": {
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "period": {
                    "type": "string"
                },
                "periodEnd": {
                    "type": "string"
                },
                "periodStart": {
                    "type": "string"
                },
                "spent": {
                    "type": "number"
                }
            }
        },
        "github_com_vasconcellos_financial-control_internal_domain_dto.CategoryResponse": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "parentId": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "github_com_vasconcellos_financial-control_internal_domain_dto.CreateBudgetRequest": {
            "type": "object",
            "required": [
                "alertPercent",
                "amount",
                "categoryId",
                "currency",
                "period",
                "periodEnd",
                "periodStart"
            ],
            "properties": {
                "alertPercent": {
                    "type": "number"
                },
                "amount": {
                    "type": "number"
                },
                "categoryId": {
                    "type": "string"
                },
                "currency": {
                    "type": "string",
                    "enum": [
                        "USD",
                        "EUR",
                        "CHF",
                        "GBP",
                        "BRL"
                    ]
                },
                "period": {
                    "type": "string",
                    "enum": [
                        "monthly",
                        "quarterly",
                        "yearly"
                    ]
                },
                "periodEnd": {
                    "type": "string"
                },
                "periodStart": {
                    "type": "string"
                }
            }
        },
        "github_com_vasconcellos_financial-control_internal_domain_dto.CreateCategoryRequest": {
            "type": "object",
            "required": [
                "name",
                "type"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "parentId": {
                    "type": "string"
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "income",
                        "expense"
                    ]
                }
            }
        },
        "github_com_vasconcellos_financial-control_internal_domain_dto.CreateGoalRequest": {
            "type": "object",
            "required": [
                "currency",
                "deadline",
                "name",
                "targetAmount"
            ],
            "properties": {
                "currency": {
                    "type": "string",
                    "enum": [
                        "USD",
                        "EUR",
                        "CHF",
                        "GBP",
                        "BRL"
                    ]
                },
                "deadline": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "targetAmount": {
                    "type": "number"
                }
            }
        },
...
        "github_com_vasconcellos_financial-control_internal_domain_dto.GoalResponse": {
            "type": "object",
            "properties": {
                "currency": {
                    "type": "string"
                },
                "currentAmount": {
                    "type": "number"
                },
                "deadline": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "targetAmount": {
                    "type": "number"
                }
            }
        },
...
        "internal_interfaces_http_handler.ErrorResponse": {
            "description": "Standard error response structure",
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "error message"
                }
            }
        },
...
    }
}